laird: set datarate based off spreading factor
In the channels plans the datarate is unspecified however spreading
factor is. We take the SF and map it to the DR bit mask.
Note that 'datarate' is not the raw data rate to use but a bit field.

Bug 11752

+++ util/config.go
@@ -18,7 +18,7 @@ type ChannelConf struct {
 	IfValue      int32   `json:"if"`
 	Bandwidth    *uint32 `json:"bandwidth,omitempty"`
 	Datarate     *uint32 `json:"datarate,omitempty"`
-	SpreadFactor *uint8  `json:"spread_factor,omitempty"`
+	SpreadFactor *uint32 `json:"spread_factor,omitempty"`
 }
 
 type ChannelFreqConf struct {
